requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pyswisseph==2.10.*
pydantic==2.*
